Add a 'Silhouette' render mode. It's pretty janky and very slow, and results in some bad API changes. TODO move the logic to the shader.
The slowness and API badness can be fixed by moving this work to the shader, so we can work post-object-transform, and in parallel.
We'll need to pass the face normals/centroids, and line adjacency list.
I'm committing this since the behavior is pretty much right (although it's not accurate for concave meshes).